+++ mods/src/creative/creative.cpp
@@ -44,6 +44,9 @@ static int32_t Inventory_setupDefault_FillingContainer_addItem_call_injection(un
     inventory_add_item(filling_container, *Tile_topSnow, true);
     inventory_add_item(filling_container, *Tile_ice, true);
     inventory_add_item(filling_container, *Tile_invisible_bedrock, true);
+    inventory_add_item(filling_container, *Tile_bedrock, true);
+    inventory_add_item(filling_container, *Tile_info_updateGame1, true);
+    inventory_add_item(filling_container, *Tile_info_updateGame2, true);
 
     return ret;
 }